GW-98: Refactored navigation bar
