+++ client/src/shared/components/navigation-bar/NavigationBar.jsx
@@ -1,16 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { useMediaQuery } from 'react-responsive';
-import { Link, Outlet, useLocation } from 'react-router-dom';
+import { Outlet, useLocation } from 'react-router-dom';
 import { Box } from '@mui/material';
-import { ReactComponent as NavLogoBig } from 'assets/gisau-logo/gisau_white.svg';
-import { ReactComponent as NavLogoSmall } from 'assets/gisau-logo/gisau_white_small.svg';
-import gamesIconLeft from 'assets/games/controllerL.png';
-import gamesIconRight from 'assets/games/controllerR.png';
-import gamesIconLRed from 'assets/games/controllerLred.png';
-import gamesIconRRed from 'assets/games/controllerRred.png';
-import Dashboard from 'pages/games/Dashboard';
 import pages from './constants';
-import MenuInterface from './MenuInterface';
+import MobileNavBar from './components/MobileNavBar';
+import NavBarLogo from './components/NavBarLogo';
+import DesktopNavBar from './components/DesktopNavBar';
 
 export default function NavigationBar() {
 	const location = useLocation();
@@ -21,26 +16,6 @@ export default function NavigationBar() {
 	const isMobile = useMediaQuery({ query: '(max-width: 767px)' });
 
 	const [isMenuOpen, setIsMenuOpen] = useState(false);
-
-	const handleClosingMenu = () => {
-		setIsMenuOpen(false);
-
-		// Unsets Background Scrolling to use when SideDrawer/Modal is closed
-		document.body.style.overflow = 'unset';
-	};
-
-	const handleOpeningMenu = () => {
-		setIsMenuOpen(true);
-
-		// Disables Background Scrolling whilst the SideDrawer/Modal is open
-		if (typeof window !== 'undefined' && window.document) {
-			document.body.style.overflow = 'hidden';
-		}
-	};
-
-	useEffect(() => {
-		handleClosingMenu(); // Close the navigation panel
-	}, [location.pathname]);
 
 	useEffect(() => {
 		const handleScroll = () => {
@@ -59,104 +34,28 @@ export default function NavigationBar() {
 			window.removeEventListener('scroll', handleScroll);
 		};
 	}, []);
-
-	const ScrollToTop = () => {
-		window.scrollTo(0, 0);
-	};
-
 	const isGamesPage = location.pathname === '/dashboard';
 
 	return isGamesPage ? (
-		<Dashboard />
+		<Outlet />
 	) : (
 		<div className={isMenuOpen ? 'overflow-y-hidden' : 'overflow-y-visible'}>
 			<Box className='flex justify-between items-center z-30 w-full mt-[5vh] absolute'>
-				<Box className='ml-6 sm:ml-20 md:ml-6 lg:ml-20'>
-					<Link to='/'>
-						{hasLandingImage ? (
-							<NavLogoBig className='w-14 sm:w-16 h-auto' />
-						) : (
-							<NavLogoSmall className='w-14 sm:w-16 h-auto' />
-						)}
-					</Link>
-				</Box>
+				<NavBarLogo hasLandingImage={hasLandingImage} />
 				{isMobile ? (
-					<div className='mr-6 sm:mr-20'>
-						<MenuInterface
-							isOpen={isMenuOpen}
-							closeHandler={handleClosingMenu}
-							openHandler={handleOpeningMenu}
-							hasLandingImage={hasLandingImage}
-						/>
-					</div>
+					<MobileNavBar
+						isMenuOpen={isMenuOpen}
+						setIsMenuOpen={setIsMenuOpen}
+						hasLandingImage={hasLandingImage}
+						location={location}
+					/>
 				) : (
-					<Box className='right-0 fixed'>
-						<Box
-							className={`flex mr-6 lg:mr-20 navbar ${
-								hasLandingImage ? bgColor : 'bg-white bg-opacity-70'
-							} h-14 rounded-[15px]`}
-						>
-							{pages.map((page) => (
-								<Link
-									key={page.name}
-									to={page.path}
-									className='px-5 pt-3'
-									onClick={ScrollToTop}
-									target={page.name === 'Games' ? '_blank' : '_self'}
-								>
-									{page.name === 'Games' ? (
-										<div className='flex items-center ml-[-10] self-center'>
-											{hasLandingImage ? (
-												<img
-													src={gamesIconLeft}
-													alt='Games'
-													className='image-class'
-													style={{ height: 'auto', width: '20px' }}
-												/>
-											) : (
-												<img
-													src={gamesIconLRed}
-													alt='Red Left'
-													className='image-class'
-													style={{ height: 'auto', width: '20px' }}
-												/>
-											)}
-											<p
-												className={`underline-animation font-oswald text-xl 
-										${hasLandingImage ? 'text-white' : 'text-primary underline-animation-red'}
-										${page.path === location.pathname && 'underlined'}`}
-											>
-												{page.name}
-											</p>
-											{hasLandingImage ? (
-												<img
-													src={gamesIconRight}
-													alt='Games'
-													className='image-class'
-													style={{ height: 'auto', width: '20px' }}
-												/>
-											) : (
-												<img
-													src={gamesIconRRed}
-													alt='Red Left'
-													className='image-class'
-													style={{ height: 'auto', width: '20px' }}
-												/>
-											)}
-										</div>
-									) : (
-										<p
-											className={`underline-animation font-oswald text-xl 
-									  ${hasLandingImage ? 'text-white' : 'text-primary underline-animation-red'}
-									  ${page.path === location.pathname && 'underlined'}`}
-										>
-											{page.name}
-										</p>
-									)}
-								</Link>
-							))}
-						</Box>
-					</Box>
+					<DesktopNavBar
+						bgColor={bgColor}
+						hasLandingImage={hasLandingImage}
+						location={location}
+						pages={pages}
+					/>
 				)}
 			</Box>
 			<Box>
